reorganizacao dos diretorios e codigo rodando redondo
