+++ slicing/Altprint/slicer.py
@@ -1,7 +1,7 @@
 from abc import ABC, abstractmethod
 import trimesh
 from shapely.geometry import MultiPolygon
-from altprint.height_method import HeightMethod
+from Altprint.height_method import HeightMethod
 
 # este arquivo tem como funcionalidade, ler um arquivo em stl, movimentar a peça e depois fatiar ela em planos associados a cada camada (STLslicer), isso é armazenado em um objeto que recebe essas informações e armazena a altura de cada camada, a geometria de cada camada e os limites superior e inferior da peça (SlicedPlanes)
