+++ slicing/Altprint/standart.py
