+++ slicing/Altprint/standart_copy_copy.py
