+++ slicing/batch_flex_slicing.py
@@ -1,20 +1,9 @@
-from altprint.printable.flex_continuous import FlexProcess, FlexPrint
-from altprint.printable.multi import MultiPrint, MultiProcess
+from Altprint.flex_continuous import FlexProcess, FlexPrint
+from Altprint.multi import MultiPrint, MultiProcess
 
-import os
-
-# Para Ricas **
-os.chdir("D:/códigosVScode/continuous_altprint/Altprint")
-
-# Para Stavo notebook
-# os.chdir("C:/Users/gusta/OneDrive/Documentos/GitHub/continuous_altprint/Altprint")
-
-# Para Stavo PC
-# os.chdir("C:/Users/Gustavo Mariano/Documents/Github/continuous_altprint/Altprint")
-
-process1 = FlexProcess(settings_file='flex_parameters.yml')
-process2 = FlexProcess(settings_file='flex_parameters2.yml')
-process3 = FlexProcess(settings_file='flex_parameters3.yml')
+process1 = FlexProcess(settings_file="slicing/parameters/flex_parameters.yml")
+process2 = FlexProcess(settings_file="slicing/parameters/flex_parameters2.yml")
+process3 = FlexProcess(settings_file="slicing/parameters/flex_parameters3.yml")
 
 part1 = FlexPrint(process1)
 part1.slice()
@@ -30,10 +19,10 @@ part3.make_layers()
 
 multi_process = MultiProcess(
     parts=[part1, part2, part3],
-    start_script="scripts/start_prusaMK3S.gcode",
-    end_script="scripts/end_prusaMK3S.gcode")
+    start_script="slicing/gcode/start_prusaMK3S.gcode",
+    end_script="slicing/gcode/end_prusaMK3S.gcode")
 
 multi_part = MultiPrint(multi_process)
 multi_part.slice()
 multi_part.make_layers()
-multi_part.export_gcode("sliced_geometry.gcode")
+multi_part.export_gcode("slicing/parameters/sliced_geometry.gcode")
