+++ slicing/batch_standart_slicing.py
@@ -1,20 +1,10 @@
-from altprint.printable.standart_infillSlicer import StandartProcess, StandartPrint
-from altprint.printable.multi import MultiPrint, MultiProcess
-
-import os
-
-# Para Ricas **
-os.chdir("D:/códigosVScode/continuous_altprint/Altprint")
-
-# Para Stavo notebook
-# os.chdir("C:/Users/gusta/OneDrive/Documentos/GitHub/continuous_altprint/Altprint")
-
-# Para Stavo PC
-# os.chdir("C:/Users/Gustavo Mariano/Documents/Github/continuous_altprint/Altprint")
+from Altprint.standart_infillSlicer import StandartProcess, StandartPrint
+from Altprint.multi import MultiPrint, MultiProcess
 
 # --------------------------------------------------------------------
 
-process1 = StandartProcess(settings_file='standart_parameters.yml')
+process1 = StandartProcess(
+    settings_file='slicing/parameters/standart_parameters.yml')
 part1 = StandartPrint(process1)
 
 
@@ -23,7 +13,8 @@ part1.make_layers()
 
 # --------------------------------------------------------------------
 
-process2 = StandartProcess(settings_file='standart_parameters2.yml')
+process2 = StandartProcess(
+    settings_file='slicing/parameters/standart_parameters2.yml')
 part2 = StandartPrint(process2)
 
 
@@ -32,7 +23,8 @@ part2.make_layers()
 
 # --------------------------------------------------------------------
 
-process3 = StandartProcess(settings_file='standart_parameters3.yml')
+process3 = StandartProcess(
+    settings_file='slicing/parameters/standart_parameters3.yml')
 part3 = StandartPrint(process3)
 
 
@@ -43,10 +35,10 @@ part3.make_layers()
 
 multi_process = MultiProcess(
     parts=[part1, part2, part3],
-    start_script="scripts/start_prusaMK3S.gcode",
-    end_script="scripts/end_prusaMK3S.gcode")
+    start_script="slicing/gcode/start_prusaMK3S.gcode",
+    end_script="slicing/gcode/end_prusaMK3S.gcode")
 
 multi_part = MultiPrint(multi_process)
 multi_part.slice()
 multi_part.make_layers()
-multi_part.export_gcode("sliced_geometry.gcode")
+multi_part.export_gcode("slicing/parameters/sliced_geometry.gcode")
