+++ slicing/flex_slicing.py
@@ -1,20 +1,12 @@
-from altprint.printable.flex_continuous import FlexProcess, FlexPrint
 
-import os
 
-# Para Ricas **
-os.chdir("D:/códigosVScode/continuous_altprint/Altprint")
+from Altprint.flex_continuous import FlexProcess, FlexPrint
 
-# Para Stavo notebook
-# os.chdir("C:/Users/gusta/OneDrive/Documentos/GitHub/continuous_altprint/Altprint")
 
-# Para Stavo PC
-# os.chdir("C:/Users/Gustavo Mariano/Documents/Github/continuous_altprint/Altprint")
-
-process = FlexProcess(settings_file='flex_parameters.yml')
+process = FlexProcess(settings_file="slicing/parameters/flex_parameters.yml")
 part = FlexPrint(process)
 
 
 part.slice()
 part.make_layers()
-part.export_gcode("sliced_geometry.gcode")
+part.export_gcode("slicing/parameters/sliced_geometry.gcode")
